requirements.txt
# requirements.txt
# Data Handling
pandas>=1.5.0

# AWS S3
boto3>=1.26.0

# MongoDB
pymongo>=4.3.0
certifi>=2023.7.22

# Environment Variables
python-dotenv>=1.0.0

# Korean NLP
konlpy>=0.6.0
JPype1>=1.4.0

# Text Processing
emoji>=2.0.0
regex>=2022.10.31

# Logging & Utilities
tqdm>=4.64.0

# Testing
pytest>=7.2.0
